merge(clara): integração das telas mobile
