+++ app/src/main/java/com/yannk/respira/ui/components/VectorImg.kt
@@ -1,0 +1,18 @@
+package com.yannk.respira.ui.components
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+
+@Composable
+fun VectorImg(source: Int, modifier: Modifier = Modifier) {
+    Image(
+        painter = painterResource(id = source),
+        contentDescription = null,
+        modifier = modifier.fillMaxWidth(),
+        contentScale = ContentScale.FillWidth
+    )
+}
