+++ app/src/main/java/com/yannk/respira/ui/presentation/WelcomeScreen.kt
@@ -1,4 +1,4 @@
-package com.yannk.respira.presentation
+package com.yannk.respira.ui.presentation
 
 
 import androidx.compose.foundation.Image
@@ -11,14 +11,13 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.yannk.respira.R
-import com.yannk.respira.components.FundoImg
-import com.yannk.respira.components.VectorImg
+import com.yannk.respira.ui.components.FundoImg
+import com.yannk.respira.ui.components.VectorImg
 
 
 @Preview(showBackground = false)
